Move fetch into a custom hook
